refactoring widget, logic work in progress
