+++ lib/common_widgets/gcw_custom_date_picker.dart
@@ -1,29 +1,35 @@
 import 'package:flutter/material.dart';
+import 'package:gc_wizard/common_widgets/dropdowns/gcw_dropdown.dart';
+import 'package:gc_wizard/common_widgets/spinners/gcw_dropdown_spinner.dart';
 import 'package:gc_wizard/common_widgets/spinners/gcw_integer_spinner.dart';
 import 'package:gc_wizard/common_widgets/spinners/spinner_constants.dart';
+import 'package:gc_wizard/tools/science_and_technology/date_and_time/calendar/logic/calendar.dart';
+import 'package:gc_wizard/tools/science_and_technology/date_and_time/calendar/logic/calendar_constants.dart';
 
-class GCWDatePicker extends StatefulWidget {
-  final void Function(DateTime) onChanged;
-  final DateTime date;
+class GCWCustomDatePicker extends StatefulWidget {
+  final void Function(CustomCalendarDate) onChanged;
+  final CustomCalendarDate date;
+  final CalendarSystem type;
 
   final TextEditingController? yearController;
   final TextEditingController? monthController;
   final TextEditingController? dayController;
 
-  const GCWDatePicker({
+  const GCWCustomDatePicker({
     super.key,
     required this.onChanged,
     required this.date,
+    this.type = CalendarSystem.GREGORIANCALENDAR,
     this.yearController,
     this.monthController,
     this.dayController,
   });
 
   @override
-  _GCWDatePickerState createState() => _GCWDatePickerState();
+  _GCWCustomDatePickerState createState() => _GCWCustomDatePickerState();
 }
 
-class _GCWDatePickerState extends State<GCWDatePicker> {
+class _GCWCustomDatePickerState extends State<GCWCustomDatePicker> {
   late int _currentYear;
   late int _currentMonth;
   late int _currentDay;
@@ -39,13 +45,13 @@ class _GCWDatePickerState extends State<GCWDatePicker> {
   }
 
   @override
-  void didUpdateWidget(GCWDatePicker oldWidget) {
+  void didUpdateWidget(GCWCustomDatePicker oldWidget) {
     super.didUpdateWidget(oldWidget);
     initValues();
   }
 
   void initValues() {
-    DateTime date = widget.date;
+    CustomCalendarDate date = widget.date;
     _currentYear = date.year;
     _currentMonth = date.month;
     _currentDay = date.day;
@@ -86,27 +92,32 @@ class _GCWDatePickerState extends State<GCWDatePicker> {
               )),
         ),
         Expanded(
-            child: Padding(padding: const EdgeInsets.only(left: 2, right: 2), child: _buildMonthSpinner())),
+            child: Padding(
+                padding: const EdgeInsets.only(left: 2, right: 2),
+                child: _buildMonthSpinner(widget.type)
+            )),
         Expanded(
             child: Padding(
           padding: const EdgeInsets.only(left: 2),
-          child: _buildDaySpinner(),
+          child: _buildDaySpinner(widget.type),
         ))
       ],
     );
   }
 
-  Widget _buildDaySpinner() {
+  Widget _buildDaySpinner(CalendarSystem type) {
+    int maxDays = 31;
+    if (type == CalendarSystem.POTRZEBIECALENDAR) maxDays = 10;
+
     return GCWIntegerSpinner(
       focusNode: _dayFocusNode,
       layout: SpinnerLayout.VERTICAL,
       controller: widget.dayController,
       value: _currentDay,
       min: 1,
-      max: 31,
+      max: maxDays,
       onChanged: (value) {
         setState(() {
-
           _currentDay = value;
           _setCurrentValueAndEmitOnChange();
         });
@@ -114,7 +125,32 @@ class _GCWDatePickerState extends State<GCWDatePicker> {
     );
   }
 
-  Widget _buildMonthSpinner() {
+  Widget _buildMonthSpinner(CalendarSystem type) {
+    switch (type) {
+      case CalendarSystem.ISLAMICCALENDAR:
+      case CalendarSystem.PERSIANYAZDEGARDCALENDAR:
+      case CalendarSystem.HEBREWCALENDAR:
+      case CalendarSystem.POTRZEBIECALENDAR:
+      case CalendarSystem.COPTICCALENDAR:
+      return GCWDropDownSpinner(
+        index: _currentMonth,
+        layout: SpinnerLayout.VERTICAL,
+        items: MONTH_NAMES[type]!.entries.map((entry) {
+          return GCWDropDownMenuItem(value: entry.key, child: entry.value);
+        }).toList(),
+        onChanged: (value) {
+          setState(() {
+            _currentMonth = value + 1;
+            print(_currentMonth);
+            _setCurrentValueAndEmitOnChange();
+            if (_currentMonth.toString().length == 2) {
+              FocusScope.of(context).requestFocus(_dayFocusNode);
+            }
+          });
+        },
+      );
+      case CalendarSystem.JULIANCALENDAR:
+      case CalendarSystem.GREGORIANCALENDAR:
       return GCWIntegerSpinner(
         focusNode: _monthFocusNode,
         layout: SpinnerLayout.VERTICAL,
@@ -133,9 +169,12 @@ class _GCWDatePickerState extends State<GCWDatePicker> {
           });
         },
       );
+      default: return Container();
+    }
   }
 
   void _setCurrentValueAndEmitOnChange() {
-    widget.onChanged(DateTime(_currentYear, _currentMonth, _currentDay));
+    widget.onChanged(CustomCalendarDate(
+        year: _currentYear, month: _currentMonth, day: _currentDay));
   }
 }
